update [backend] DEL POST PUT authorize event same as student's email

+++ BackEnd/src/main/java/sit/int221/eventsservice/controllers/EventController.java
@@ -4,10 +4,14 @@ import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import sit.int221.eventsservice.advice.HandleExceptionForbidden;
@@ -16,7 +20,10 @@ import sit.int221.eventsservice.dtos.Event.EventPutDTO;
 import sit.int221.eventsservice.dtos.Event.EventDTO;
 import sit.int221.eventsservice.entities.Event;
 import sit.int221.eventsservice.entities.Category;
+import sit.int221.eventsservice.entities.Role;
+import sit.int221.eventsservice.entities.User;
 import sit.int221.eventsservice.repositories.EventRepository;
+import sit.int221.eventsservice.repositories.UserRepository;
 import sit.int221.eventsservice.services.EventService;
 
 import javax.validation.Valid;
@@ -30,10 +37,12 @@ public class EventController {
     private ModelMapper modelMapper;
     @Autowired
     private EventRepository repository;
+    @Autowired
+    private UserRepository userRepository;
 
     @GetMapping({"/{Id}"})
     public EventDTO getEventById(@PathVariable Integer Id) throws HandleExceptionForbidden {
-        return this.eventService.getSimpleEventById(Id);
+        return this.eventService.getEventById(Id);
     }
 
     @GetMapping({""})
@@ -42,11 +51,26 @@ public class EventController {
     }
 
     @DeleteMapping({"/{Id}"})
-    public void delete(@PathVariable Integer Id) {
-        repository.findById(Id).orElseThrow(() ->
-                new ResponseStatusException(HttpStatus.NOT_FOUND,
-                        Id + " does not exist !!!"));
-        repository.deleteById(Id);
+    public void delete(@PathVariable Integer Id) throws HandleExceptionForbidden {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User userLogin = userRepository.findByEmail(auth.getPrincipal().toString());
+        if (userLogin.getRole().equals(Role.admin)) {
+            repository.findById(Id).orElseThrow(() ->
+                    new ResponseStatusException(HttpStatus.NOT_FOUND,
+                            Id + " does not exist !!!"));
+            repository.deleteById(Id);
+        } else if (userLogin.getRole().equals(Role.student)) {
+            Event event = repository.findById(Id).orElseThrow(() ->
+                    new ResponseStatusException(HttpStatus.NOT_FOUND,
+                            Id + " does not exist !!!"));
+            if (Objects.equals(event.getUser().getEmail(), userLogin.getEmail())) {
+                repository.deleteById(Id);
+            } else {
+                throw new HandleExceptionForbidden("You are not owner of this event");
+            }
+        } else {
+            throw new HandleExceptionForbidden("You are not owner of this event");
+        }
     }
 
     @PostMapping({""})
@@ -55,29 +79,67 @@ public class EventController {
     }
 
     @PutMapping({"/{Id}"})
-    public ResponseEntity update(@Valid @RequestBody EventPutDTO updateEvent, @PathVariable Integer Id) throws OverlappedExceptionHandler {
+    public ResponseEntity update(@Valid @RequestBody EventPutDTO updateEvent, @PathVariable Integer Id) throws OverlappedExceptionHandler, HandleExceptionForbidden {
         Date newEventStartTime = Date.from(updateEvent.getEventStartTime());
         Date newEventEndTime = eventService.findEndDate(Date.from(updateEvent.getEventStartTime()), updateEvent.getEventDuration());
         List<EventDTO> eventList = getEvents();
-        for (int i = 0; i < eventList.size(); i++) {
-            if (updateEvent.getEventCategory().getId() == eventList.get(i).getEventCategory().getId()&& eventList.get(i).getId() != Id) { //เช็คเฉพาะ EventCategory เดียวกัน และถ้าอัพเดตตัวเดิมไม่ต้องเช็ค overlapped
-                Date eventStartTime = Date.from(eventList.get(i).getEventStartTime());
-                Date eventEndTime = eventService.findEndDate(Date.from(eventList.get(i).getEventStartTime()), eventList.get(i).getEventDuration());
-                if (newEventStartTime.before(eventStartTime) && newEventEndTime.after(eventStartTime) ||
-                        newEventStartTime.before(eventEndTime) && newEventEndTime.after(eventEndTime) ||
-                        newEventStartTime.before(eventStartTime) && newEventEndTime.after(eventEndTime) ||
-                        newEventStartTime.after(eventStartTime) && newEventEndTime.before(eventEndTime) ||
-                        newEventStartTime.equals(eventStartTime)) {
-                    throw new OverlappedExceptionHandler("Time is Overlapped");
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User userLogin = userRepository.findByEmail(auth.getPrincipal().toString());
+
+        if (userLogin.getRole().equals(Role.admin) || userLogin.getRole().equals(Role.lecturer)) {
+            for (EventDTO eventDTO : eventList) {
+                if (Objects.equals(updateEvent.getEventCategory().getId(), eventDTO.getEventCategory().getId()) && eventDTO.getId() != Id) { //เช็คเฉพาะ EventCategory เดียวกัน และถ้าอัพเดตตัวเดิมไม่ต้องเช็ค overlapped
+                    Date eventStartTime = Date.from(eventDTO.getEventStartTime());
+                    Date eventEndTime = eventService.findEndDate(Date.from(eventDTO.getEventStartTime()), eventDTO.getEventDuration());
+                    if (newEventStartTime.before(eventStartTime) && newEventEndTime.after(eventStartTime) ||
+                            newEventStartTime.before(eventEndTime) && newEventEndTime.after(eventEndTime) ||
+                            newEventStartTime.before(eventStartTime) && newEventEndTime.after(eventEndTime) ||
+                            newEventStartTime.after(eventStartTime) && newEventEndTime.before(eventEndTime) ||
+                            newEventStartTime.equals(eventStartTime)) {
+                        throw new OverlappedExceptionHandler("Time is Overlapped");
+                    }
                 }
             }
+            Event event = repository.findById(Id).orElseThrow(
+                    () -> new ResponseStatusException(HttpStatus.BAD_REQUEST)
+            );
+            modelMapper.map(updateEvent, event);
+            repository.saveAndFlush(event);
+            return ResponseEntity.status(200).body(event);
+        } else if (userLogin.getRole().equals(Role.student)) {
+            Event eventForCheck =  repository.findById(Id).orElseThrow(
+                    () -> new ResponseStatusException(HttpStatus.BAD_REQUEST)
+            );
+            if (Objects.equals(updateEvent.getBookingEmail(), userLogin.getEmail())) {
+                if(Objects.equals(updateEvent.getBookingEmail(), eventForCheck.getBookingEmail())) {
+                    for (EventDTO eventDTO : eventList) {
+                        if (Objects.equals(updateEvent.getEventCategory().getId(), eventDTO.getEventCategory().getId()) && eventDTO.getId() != Id) { //เช็คเฉพาะ EventCategory เดียวกัน และถ้าอัพเดตตัวเดิมไม่ต้องเช็ค overlapped
+                            Date eventStartTime = Date.from(eventDTO.getEventStartTime());
+                            Date eventEndTime = eventService.findEndDate(Date.from(eventDTO.getEventStartTime()), eventDTO.getEventDuration());
+                            if (newEventStartTime.before(eventStartTime) && newEventEndTime.after(eventStartTime) ||
+                                    newEventStartTime.before(eventEndTime) && newEventEndTime.after(eventEndTime) ||
+                                    newEventStartTime.before(eventStartTime) && newEventEndTime.after(eventEndTime) ||
+                                    newEventStartTime.after(eventStartTime) && newEventEndTime.before(eventEndTime) ||
+                                    newEventStartTime.equals(eventStartTime)) {
+                                throw new OverlappedExceptionHandler("Time is Overlapped");
+                            }
+                        }
+                    }
+                    Event event = repository.findById(Id).orElseThrow(
+                            () -> new ResponseStatusException(HttpStatus.BAD_REQUEST)
+                    );
+                    modelMapper.map(updateEvent, event);
+                    repository.saveAndFlush(event);
+                    return ResponseEntity.status(200).body(event);
+                } else {
+                    throw new HandleExceptionForbidden("You are not owner of this event");
+                }
+            } else {
+                throw new HandleExceptionForbidden("The booking email must be the same as the student's email");
+            }
+        } else {
+            throw new HandleExceptionForbidden("You are not allowed to update this event");
         }
-        Event event = repository.findById(Id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.BAD_REQUEST)
-        );
-        modelMapper.map(updateEvent, event);
-        repository.saveAndFlush(event);
-        return ResponseEntity.status(200).body(event);
     }
